Support updated OT 0.14 / tracing-OT 0.13 crate versions (#23)
* Support updated OT 0.14 / tracing-OT 0.12 crate versions

Add support for opentelemetry 0.14 and tracing-opentelemetry 0.13.
Make corresponding changes to the features, docs, and examples.
Due to the use of two different versions of the same packages, some
reorganization had to be done for the internal modules.

Addresses issue #22.

* Re-export versioned set_otel_parent `fn`s for use in macro
