+++ Cargo.toml
@@ -20,7 +20,8 @@ categories = ["asynchronous", "web-programming"]
 
 [features]
 default = ["opentelemetry_0_13", "emit_event_on_error"]
-opentelemetry_0_13 = ["opentelemetry", "tracing-opentelemetry"]
+opentelemetry_0_13 = ["opentelemetry_0_13_pkg", "tracing-opentelemetry_0_12_pkg"]
+opentelemetry_0_14 = ["opentelemetry_0_14_pkg", "tracing-opentelemetry_0_13_pkg"]
 emit_event_on_error = []
 
 [dependencies]
@@ -31,8 +32,10 @@ tracing = "0.1.19"
 tracing-futures = "0.2.4"
 futures = "0.3.5"
 uuid = { version = "0.8.1", features = ["v4"] }
-opentelemetry = { version = "0.13", optional = true }
-tracing-opentelemetry = { version = "0.12", optional = true }
+opentelemetry_0_13_pkg = { package = "opentelemetry", version = "0.13", optional = true }
+opentelemetry_0_14_pkg = { package = "opentelemetry", version = "0.14", optional = true }
+tracing-opentelemetry_0_12_pkg = { package = "tracing-opentelemetry",version = "0.12", optional = true }
+tracing-opentelemetry_0_13_pkg = { package = "tracing-opentelemetry", version = "0.13", optional = true }
 
 [dev-dependencies]
 tracing-subscriber = { version = "0.2.12", features = ["registry", "env-filter"] }
