+++ examples/opentelemetry/Cargo.toml
@@ -9,9 +9,9 @@ license = "MIT/Apache-2.0"
 [dependencies]
 actix-web = "=4.0.0-beta.7"
 tracing = "0.1.19"
-opentelemetry = { version = "0.13", features = ["rt-tokio-current-thread"] }
-opentelemetry-jaeger = { version = "0.12", features = ["tokio"] }
-tracing-opentelemetry = { version = "0.12" }
+opentelemetry = { version = "0.14", features = ["rt-tokio-current-thread"] }
+opentelemetry-jaeger = { version = "0.13", features = ["tokio"] }
+tracing-opentelemetry = { version = "0.13" }
 tracing-subscriber = { version = "0.2.12", features = ["registry", "env-filter"] }
 tracing-bunyan-formatter = "0.1.6"
-tracing-actix-web = { path = "../.." }
+tracing-actix-web = { path = "../..", features = ["opentelemetry_0_14"] }
